refactor: Use sql-formatter

+++ db/queries/bank.sql
@@ -5,30 +5,41 @@ SELECT * FROM "Bank" WHERE id = $1;
 SELECT * FROM "Bank" WHERE "userId" = $1 ORDER BY "createdAt", "accountName";
 
 -- name: GetSystemBankById :one
-SELECT b.* FROM "Bank" b JOIN "User" u ON b."userId" = u.id WHERE u.role = 'SYSTEM'::"Role" AND b.id = $1;
+SELECT b.* FROM "Bank" b JOIN "User" u ON b."userId" = u.id WHERE u.role = 'SYSTEM' AND b.id = $1;
 
 -- name: GetSystemBanks :many
-SELECT b.* FROM "Bank" b JOIN "User" u ON b."userId" = u.id WHERE u.role = 'SYSTEM'::"Role" ORDER BY b."createdAt", b."accountName";
+SELECT b.* FROM "Bank" b JOIN "User" u ON b."userId" = u.id WHERE u.role = 'SYSTEM' ORDER BY b."createdAt", b."accountName";
 
 -- name: GetSystemBanksByBankName :many
-SELECT b.* FROM "Bank" b JOIN "User" u ON b."userId" = u.id WHERE u.role = 'SYSTEM'::"Role" AND b.name = $1;
-
-
+SELECT b.* FROM "Bank" b JOIN "User" u ON b."userId" = u.id WHERE u.role = 'SYSTEM' AND b.name = $1;
 
 -- name: CreateBank :one
-INSERT INTO "Bank" (id, "userId", name, "accountName", "accountNumber", "updatedAt") 
-VALUES (gen_random_uuid(), $1, $2, $3, $4, now())
-RETURNING *;
+INSERT INTO
+	"Bank" (id, "userId", name, "accountName", "accountNumber", "updatedAt")
+VALUES
+	(gen_random_uuid(), $1, $2, $3, $4, now())
+RETURNING
+	*;
 
 -- name: CreateSystemBank :exec
-INSERT INTO "Bank" (id, "userId", name, "accountName", "accountNumber", "updatedAt")
-SELECT gen_random_uuid(), id, $1, $2, $3, now() FROM "User" WHERE role = 'SYSTEM'::"Role" LIMIT 1;
-
-
+INSERT INTO
+	"Bank" (id, "userId", name, "accountName", "accountNumber", "updatedAt")
+SELECT
+	gen_random_uuid(),
+	id,
+	$1,
+	$2,
+	$3,
+	now()
+FROM
+	"User"
+WHERE
+	role = 'SYSTEM'
+LIMIT
+	1;
 
 -- name: UpdateBank :exec
-UPDATE
-	"Bank"
+UPDATE "Bank"
 SET
 	name = $2,
 	"accountName" = COALESCE(sqlc.narg('accountName'), "accountName"),
@@ -38,8 +49,7 @@ WHERE
 	id = $1;
 
 -- name: UpdateSystemBank :exec
-UPDATE
-	"Bank"
+UPDATE "Bank"
 SET
 	"accountName" = COALESCE(sqlc.narg('accountName'), "accountName"),
 	"accountNumber" = COALESCE(sqlc.narg('accountNumber'), "accountNumber"),
@@ -48,10 +58,8 @@ SET
 WHERE
 	id = $1;
 
-
-
 -- name: DeleteBankById :exec
-DELETE FROM "Bank" b USING "User" u WHERE b."userId" = u.id	AND b.id = $1;
+DELETE FROM "Bank" b USING "User" u WHERE b."userId" = u.id AND b.id = $1;
 
 -- name: DeleteSystemBankById :exec
-DELETE FROM "Bank" b USING "User" u WHERE b."userId" = u.id AND b.id = $1 AND u.role = 'SYSTEM'::"Role";
+DELETE FROM "Bank" b USING "User" u WHERE b."userId" = u.id AND b.id = $1 AND u.role = 'SYSTEM';
